requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
openpyxl>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
adjustText>=0.7.3
